Move clipboard helper from crypto_wallet.util into ui.base library
