+++ android/java/org/chromium/chrome/browser/crypto_wallet/util/Utils.java
@@ -5,12 +5,8 @@
 
 package org.chromium.chrome.browser.crypto_wallet.util;
 
-import static android.content.ClipDescription.MIMETYPE_TEXT_PLAIN;
-
 import android.annotation.SuppressLint;
 import android.app.Activity;
-import android.content.ClipData;
-import android.content.ClipboardManager;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -45,15 +41,11 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import org.chromium.base.ApiCompatibilityUtils;
-import org.chromium.base.BraveReflectionUtil;
 import org.chromium.base.Callbacks;
 import org.chromium.base.CommandLine;
-import org.chromium.base.ContextUtils;
 import org.chromium.base.Log;
 import org.chromium.base.shared_preferences.SharedPreferencesManager;
 import org.chromium.base.supplier.ObservableSupplier;
-import org.chromium.base.task.PostTask;
-import org.chromium.base.task.TaskTraits;
 import org.chromium.brave_wallet.mojom.AccountId;
 import org.chromium.brave_wallet.mojom.AccountInfo;
 import org.chromium.brave_wallet.mojom.AssetRatioService;
@@ -78,9 +70,7 @@ import org.chromium.chrome.browser.profiles.Profile;
 import org.chromium.chrome.browser.profiles.ProfileManager;
 import org.chromium.chrome.browser.tabmodel.TabModelSelector;
 import org.chromium.chrome.browser.util.TabUtils;
-import org.chromium.ui.base.Clipboard;
 import org.chromium.ui.text.ChromeClickableSpan;
-import org.chromium.ui.widget.Toast;
 
 import java.io.InputStream;
 import java.lang.ref.WeakReference;
@@ -125,8 +115,6 @@ public class Utils {
 
     public static int[] P3ACoinTypes = {CoinType.ETH, CoinType.SOL, CoinType.FIL};
 
-    private static final int CLEAR_CLIPBOARD_INTERVAL = 60000; // In milliseconds
-
     public static List<String> getRecoveryPhraseAsList(final String recoveryPhrase) {
         final String[] recoveryPhraseArray = recoveryPhrase.split(" ");
         return new ArrayList<>(Arrays.asList(recoveryPhraseArray));
@@ -138,66 +126,6 @@ public class Utils {
             recoveryPhrasesText = recoveryPhrasesText.concat(phrase).concat(" ");
         }
         return recoveryPhrasesText.trim();
-    }
-
-    /**
-     * Saves a given text to clipboard, shows a toast and clears it again after 60 seconds.
-     *
-     * @param context Context used to retrieve the clipboard service.
-     * @param textToCopy Text that will be copied to clipboard.
-     * @param textToShow String resource ID to display in the toast, or -1 to disable the toast.
-     * @param treatAsPasword {@code true} copy to the clipboard with
-     *     ClipDescription.EXTRA_IS_SENSITIVE flag and then clear the clipboard after {@link
-     *     #CLEAR_CLIPBOARD_INTERVAL}.
-     */
-    public static void saveTextToClipboard(
-            final Context context,
-            final String textToCopy,
-            @StringRes final int textToShow,
-            final boolean treatAsPasword) {
-        if (treatAsPasword) {
-            Clipboard.getInstance().setPassword(textToCopy);
-        } else {
-            Clipboard.getInstance().setText("" /* label */, textToCopy, false);
-        }
-
-        // Similar to ClipboardImpl.showToastIfNeeded
-        // Conditionally show a toast to avoid duplicate notifications in Android 13+
-        if (textToShow != -1 && Build.VERSION.SDK_INT <= Build.VERSION_CODES.S_V2) {
-            Toast.makeText(context, textToShow, Toast.LENGTH_SHORT).show();
-        }
-        if (!treatAsPasword) {
-            return;
-        }
-
-        PostTask.postDelayedTask(
-                TaskTraits.UI_DEFAULT, () -> clearClipboard(textToCopy), CLEAR_CLIPBOARD_INTERVAL);
-    }
-
-    public static String getTextFromClipboard(Context context) {
-        ClipboardManager clipboard =
-                (ClipboardManager) context.getSystemService(Context.CLIPBOARD_SERVICE);
-        ClipData clipData = clipboard.getPrimaryClip();
-        if (clipData != null
-                && clipData.getDescription().hasMimeType(MIMETYPE_TEXT_PLAIN)
-                && clipData.getItemCount() > 0) {
-            return clipData.getItemAt(0).getText().toString();
-        }
-
-        return "";
-    }
-
-    /**
-     * Clears the clipboard and replaces it with "***" if it matches a given text.
-     *
-     * @param textToCompare Text to compare that will trigger the clipboard clearing in case of a
-     *     match.
-     */
-    public static void clearClipboard(final String textToCompare) {
-        String clipboardText = getTextFromClipboard(ContextUtils.getApplicationContext());
-        if (textToCompare.equals(clipboardText)) {
-            BraveReflectionUtil.invokeMethod(Clipboard.class, Clipboard.getInstance(), "clear");
-        }
     }
 
     public static boolean shouldShowCryptoOnboarding() {
