+++ ui/android/BUILD.gn
@@ -1,0 +1,17 @@
+# Copyright (c) 2025 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import("//build/config/android/rules.gni")
+
+android_library("base_java") {
+  sources = [ "java/src/org/chromium/ui/base/BraveClipboardHelper.java" ]
+  deps = [
+    "//base:base_java",
+    "//base:tasks_java",
+    "//third_party/androidx:androidx_annotation_annotation_jvm_java",
+    "//ui/android:ui_no_recycler_view_java",
+  ]
+  resources_package = "org.chromium.ui"
+}
